requirements.txt
requests==2.25.1
pyquery==1.4.3
Flask==1.1.2
gunicorn==20.1.0
bitrix24-rest==1.1.1
python-json-logger==2.0.1